Move ImportGenesisAccountsFromSnapshotCmd to cmd/nibirud/airdrop.go
